The session storage management was decoupled from User to separate UserStorage class. It allows to define own different User class without need of copying or again-coding all the needed sessions stuff.
